+++ src/entities/calendar/config/weekDays.ts
@@ -1,0 +1,1 @@
+export const WEEK_DAYS = ['일', '월', '화', '수', '목', '금', '토'];
