+++ src/entities/calendar/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './useCalendarView.ts';
+export * from './dateUtils.ts';
