+++ src/entities/calendar/lib/useCalendarView.ts
@@ -1,24 +1,11 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
+import { useCalendarStore } from '../../../features/calendar';
 import { fetchHolidays } from '../api';
 
 export const useCalendarView = () => {
-  const [view, setView] = useState<'week' | 'month'>('month');
-  const [currentDate, setCurrentDate] = useState(new Date());
-  const [holidays, setHolidays] = useState<{ [key: string]: string }>({});
-
-  const navigate = (direction: 'prev' | 'next') => {
-    setCurrentDate((prevDate) => {
-      const newDate = new Date(prevDate);
-      if (view === 'week') {
-        newDate.setDate(newDate.getDate() + (direction === 'next' ? 7 : -7));
-      } else if (view === 'month') {
-        newDate.setDate(1); // 항상 1일로 설정
-        newDate.setMonth(newDate.getMonth() + (direction === 'next' ? 1 : -1));
-      }
-      return newDate;
-    });
-  };
+  const { view, setView, currentDate, setCurrentDate, holidays, setHolidays, navigate } =
+    useCalendarStore();
 
   useEffect(() => {
     setHolidays(fetchHolidays(currentDate));
