+++ src/entities/calendar/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './MonthView.tsx';
+export * from './WeekView.tsx';
